Fix cross-domain GetFeatureInfo on clickable layers
The jquery.indiciaMapPanel allows for clicking on WMS layers to fetch
and display feature information in a pop-up box. Typically this is a
cross-domain request from a website to a warehouse. Originally this
was supported in Drupal 6 with a proxy module but this has not been
ported to Drupal 7 so it has not been working.

In order to fix this for all users, not just Drupal users, the request
has been modified to use JSONP which has been provided by GeoServer
since version 2.3. GeoServer administrators must enable the JSONP
response for this to work.