Ensure geom refreshed on sref change
Previously if you entered an sref then edited it but had some
communications failure to the server after editing the sref, the geom
would be left as it was when you first entered an sref. This change
ensures that it is at least cleared (it would be calculated on the
server side if no geom submitted).